Use the 'service_ticket' column in AR storage
Values stored in ActiveRecord sessions are not mapped onto database
columns. They're serialized and stored in the 'data' column. But the
read_service_session_lookup method seems to be expecting the service
ticket to be stored in the 'service_ticket' column of the sessions
table -- which is never set! I have no idea how or if this was ever
working properly.