Issue when authenticate on every request is disabled with single sign off
When using single sign off the RubyCAS gem looks for the ticket passed
to the web application in the logout message. Unfortunately this will
be the latest ticket rather than the ticket the web application is most
likely using making it impossible to end the current session.

This change looks at the authenticate_on_every_request configuration
property to choose the correct experience to avoid the issue with
single sign off enabled and authenticate on every request disabled.